requirements.txt
jupyter
matplotlib
numpy
pillow
PyQt6
scipy
tensorflow[and-cuda]
types-tensorflow
